# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MUFAZ Kitchen POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu-items:
    get:
      operationId: listMenuItems
      tags: [menu-items]
      summary: List menu items
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: isAvailable
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu-items]
      summary: Create a menu item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu-items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu-items]
      summary: Get menu item by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMenuItem
      tags: [menu-items]
      summary: Update a menu item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMenuItem
      tags: [menu-items]
      summary: Soft-delete a menu item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /menu-items/{id}/toggle-availability:
    patch:
      operationId: toggleMenuItemAvailability
      tags: [menu-items]
      summary: Toggle availability of a menu item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── CUSTOMERS ──────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/dashboard-summary:
    get:
      operationId: getOrdersDashboardSummary
      tags: [orders]
      summary: Get dashboard summary (today's orders, revenue, active counts)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  # ─── INGREDIENTS ────────────────────────────────────────────────────────────
...
  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List all recipes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of recipes with ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecipeDetail"
  /recipes/menu-item/{menuItemId}:
    get:
      operationId: getRecipeByMenuItem
      tags: [recipes]
      summary: Get recipe for a menu item
      security:
        - bearerAuth: []
      parameters:
        - name: menuItemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe with ingredients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertRecipe
      tags: [recipes]
      summary: Create or update recipe for a menu item
      security:
        - bearerAuth: []
      parameters:
        - name: menuItemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertRecipeBody"
      responses:
        "200":
          description: Recipe saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetail"

  # ─── BILLS ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        name:
          type: string
        nameUr:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        descriptionUr:
          type: ["string", "null"]
        price:
          type: string
        isAvailable:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - categoryId
        - name
        - price
        - isAvailable
        - createdAt
        - updatedAt
    CreateMenuItemBody:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        nameUr:
          type: string
        description:
          type: string
        descriptionUr:
          type: string
        price:
          type: string
        isAvailable:
          type: boolean
        imageUrl:
          type: string
      required:
        - categoryId
        - name
        - price
    UpdateMenuItemBody:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        nameUr:
          type: string
        description:
          type: string
        descriptionUr:
          type: string
        price:
          type: string
        isAvailable:
          type: boolean
        imageUrl:
          type: string

    # CUSTOMERS
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: ["integer", "null"]
        itemName:
          type: string
        itemPrice:
          type: string
        quantity:
          type: integer
        subtotal:
          type: string
      required:
        - id
        - orderId
        - itemName
        - itemPrice
        - quantity
        - subtotal
    Order:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        totalAmount:
          type: string
        paymentMethod:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - totalAmount
        - createdAt
        - updatedAt
    OrderDetail:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        totalAmount:
          type: string
        paymentMethod:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required:
        - id
        - status
        - totalAmount
        - createdAt
        - updatedAt
        - items
    CreateOrderItemInput:
      type: object
      properties:
        menuItemId:
          type: integer
        itemName:
          type: string
        itemPrice:
          type: string
        quantity:
          type: integer
      required:
        - menuItemId
        - itemName
        - itemPrice
        - quantity
    CreateOrderBody:
      type: object
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        notes:
          type: string
        paymentMethod:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemInput"
      required:
        - items
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
        paymentMethod:
          type: string
      required:
        - status
    DashboardSummary:
      type: object
      properties:
        todayRevenue:
          type: string
        todayOrderCount:
          type: integer
        activeOrderCount:
          type: integer
        pendingOrderCount:
          type: integer
        lowStockCount:
          type: integer
        totalCustomers:
          type: integer
      required:
        - todayRevenue
        - todayOrderCount
        - activeOrderCount
        - pendingOrderCount
        - lowStockCount
        - totalCustomers

    # INGREDIENTS
...
    RecipeIngredientDetail:
      type: object
      properties:
        id:
          type: integer
        recipeId:
          type: integer
        ingredientId:
          type: integer
        ingredientName:
          type: string
        unit:
          type: string
        quantity:
          type: string
      required:
        - id
        - recipeId
        - ingredientId
        - ingredientName
        - unit
        - quantity
    RecipeDetail:
      type: object
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredientDetail"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - menuItemId
        - menuItemName
        - ingredients
        - createdAt
        - updatedAt
    RecipeIngredientInput:
      type: object
      properties:
        ingredientId:
          type: integer
        quantity:
          type: string
      required:
        - ingredientId
        - quantity
    UpsertRecipeBody:
      type: object
      properties:
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredientInput"
      required:
        - ingredients

    # BILLS
...