# lib/api-zod/src/generated/types/createMenuItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MUFAZ Kitchen POS API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMenuItemBody {
  categoryId: number;
  name: string;
  nameUr?: string;
  description?: string;
  descriptionUr?: string;
  price: string;
  isAvailable?: boolean;
  imageUrl?: string;
}