# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MUFAZ Kitchen POS API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  todayRevenue: string;
  todayOrderCount: number;
  activeOrderCount: number;
  pendingOrderCount: number;
  lowStockCount: number;
  totalCustomers: number;
}